rename SearchCommentsForStudentsTable as CommentsForStudentsTable
